iroh_blobs
Module api

The user facing API of the store.

This API is both for interacting with an in-process store and for interacting with a remote store via rpc calls.

The entry point for the api is the Store struct. There are several ways to obtain a Store instance: it is available via Deref from the different store implementations (e.g. MemStore and FsStore) as well as on the BlobsProtocol iroh protocol handler.

You can also connect to a remote store that is listening to rpc requests.

Modules

blobs

API to interact with a local blob store

downloader

API for downloads from multiple nodes.

remote

API for downloading blobs from a single remote node.

tags

Tags API

Structs

Store

The main entry point for the store API.

Tag

A named, persistent tag.

TempTag

An ephemeral, in-memory tag that protects content while the process is running.

Enums

Error

ExportBaoError

RequestError

Type Aliases

ExportBaoResult

RequestResult

Result