iroh_blobs
Module cli

Available on crate feature cli only.

Define blob-related commands.

Modules

tags

Define the tags subcommand.

Structs

BlobAddOptions

Options for the blob add command.

ProvideProgressState

Progress state for providing.

ProvideResponseEntry

Entry with a given name, size, and hash.

Enums

BlobCommands

Subcommands for the blob command.

BlobSource

Where the data should be read from.

BlobSourceIroh

Data source for adding data to iroh.

DeleteCommands

Possible delete subcommands.

ListCommands

Possible list subcommands.

OutputTarget

Where the data should be stored.

TicketOption

Whether to print an all-in-one ticket.

TicketOrHash

Possible outcomes of an input.

Functions

add

Adds data to iroh, either from a path or, if path is None, from STDIN.

add_with_opts

Adds a BlobSource given some BlobAddOptions.

aggregate_add_response

Combines the AddProgress outputs from a [Stream] into a single tuple.

consistency_check

Checks the consistency of the blobs on the running node, and repairs inconsistencies if instructed.

print_add_response

Prints out the add response.

show_download_progress

Displays the download progress for a given stream.

validate

Checks the validity of the blobs on the running node, and repairs anything invalid if instructed.