Trait iroh_blobs::downloader::Getter

pub trait Getter {
    type Connection;

    // Required method
    fn get(
        &mut self,
        kind: DownloadKind,
        conn: Self::Connection,
        progress_sender: BroadcastProgressSender
    ) -> BoxedLocal<Result<Stats, FailureAction>>;
}

Trait modelling performing a single request over a connection. This allows for IO-less testing.

Required Associated Types

type Connection

Type of connections the Getter requires to perform a download.

Required Methods

fn get( &mut self, kind: DownloadKind, conn: Self::Connection, progress_sender: BroadcastProgressSender ) -> BoxedLocal<Result<Stats, FailureAction>>

Return a future that performs the download using the given connection.