iroh_blobs::format::collection
Struct Collection

pub struct Collection { /* private fields */ }

A collection of blobs

Note that the format is subject to change.

Implementations

impl Collection

pub const HEADER: &'static [u8; 13] = b"CollectionV0."

The header for the collection format.

This is the start of the metadata blob.

pub fn to_blobs(&self) -> impl DoubleEndedIterator<Item = Bytes>

Convert the collection to an iterator of blobs, with the last being the root blob.

To persist the collection, write all the blobs to storage, and use the hash of the last blob as the collection hash.

pub async fn read_fsm( fsm_at_start_root: AtStartRoot, ) -> Result<(EndBlobNext, HashSeq, Collection)>

Read the collection from a get fsm.

Returns the fsm at the start of the first child blob (if any), the links array, and the collection.

pub async fn read_fsm_all( fsm_at_start_root: AtStartRoot, ) -> Result<(Collection, BTreeMap<u64, Bytes>, Stats)>

Read the collection and all itโs children from a get fsm.

Returns the collection, a map from blob offsets to bytes, and the stats.

pub async fn load(root: Hash, store: &impl SimpleStore) -> Result<Self>

Create a new collection from a hash sequence and metadata.

pub async fn load_db<D>(db: &D, root: &Hash) -> Result<Self>
where D: Map,

Load a collection from a store given a root hash

This assumes that both the links and the metadata of the collection is stored in the store. It does not require that all child blobs are stored in the store.

pub async fn store<D>(self, db: &D) -> Result<TempTag>
where D: Store,

Store a collection in a store. returns the root hash of the collection as a TempTag.

pub fn iter(&self) -> impl Iterator<Item = &(String, Hash)>

Iterate over the blobs in this collection

pub fn len(&self) -> usize

Get the number of blobs in this collection

pub fn is_empty(&self) -> bool

Check if this collection is empty

pub fn push(&mut self, name: String, hash: Hash)

Add the given blob to the collection.