Module iroh_blobs::get::db

Functions that use the iroh-blobs protocol in conjunction with a bao store.

Structs

GetStateNeedsConn

Intermediary state returned from get_to_db_in_steps for a download request that needs a connection to proceed.

Enums

BlobId

The id of a blob in a transfer

BlobInfo

Information about a the status of a blob in a store.

DownloadProgress

Progress updates for the get operation.

GetState

Output of get_to_db_in_steps.

Functions

blob_info

Get information about a blob in a store.

get_to_db

Get a blob or collection into a store.

get_to_db_in_steps

Get a blob or collection into a store, yielding if a connection is needed.

valid_ranges

Given a partial entry, get the valid ranges.