iroh_blobs::get::request
Function get_chunk_probe

pub async fn get_chunk_probe(
    connection: &Connection,
    hash: &Hash,
    chunk: ChunkNum,
) -> GetResult<Stats>

Probe for a single chunk of a blob.

This is used to check if a peer has a specific chunk. If the operation is successful, we have a strong indication that the peer had the chunk at the time of the request.

If the operation fails, either the connection failed or the peer did not have the chunk.

It is usually not very helpful to try to distinguish between these two cases.