iroh_blobs::get::request

Function get_verified_size

pub async fn get_verified_size(
    connection: &Connection,
    hash: &Hash,
) -> GetResult<(u64, Stats)>

Get the verified size of a blob from a peer.

This asks for the last chunk of the blob and validates the response. Note that this does not validate that the peer has all the data.