Module iroh_blobs::protocol::range_spec

Specifications for ranges selection in blobs and sequences of blobs.

The RangeSpec allows specifying which BAO chunks inside a single blob should be selected.

The RangeSpecSeq builds on top of this to select blob chunks in an entire collection.

Structs

NonEmptyRequestRangeSpecIter

An iterator over blobs in the sequence with a non-empty range specs.

RangeSpec

A chunk range specification as a sequence of chunk offsets.

RangeSpecSeq

A chunk range specification for a sequence of blobs.

RequestRangeSpecIter

An infinite iterator yielding RangeSpecs for each blob in a sequence.

Statics

EMPTY_RANGE_SPEC 🔒