iroh_blobs
Module provider

The server side API

Structs

EventSender

A sender for events related to blob transfers.

ResponseWriter

A helper struct that combines a quinn::SendStream with auxiliary information

TransferStats

The stats for a transfer of a collection or blob.

Enums

AddProgress

Progress updates for the add operation.

BatchAddPathProgress

Progress updates for the batch add operation.

Event

Events emitted by the provider informing about the current status.

SentStatus

Status of a send operation

Traits

CustomEventSender

Trait for sending blob events.

Functions

handle_connection

Handle a single connection.

handle_get

Handle a single get request.

read_request

Read the request from the getter.

send_blob

Send a blob to the client.