iroh_blobs::provider

Module events

Structs

ClientConnected

ConnectionClosed

EventMask

Event mask to configure which events are sent to the event handler.

EventSender

Notify

Newtype wrapper that wraps an event so that it is a distinct type for the notify variant.

RequestReceived

A new get request was received from the provider.

RequestTracker

Throttle

Request to throttle sending for a specific request.

TransferAborted

TransferCompleted

TransferProgress

TransferStarted

Enums

AbortReason

ConnectMode

Mode for connect events.

ObserveMode

Request mode for observe requests.

ProgressError

Errors that can occur when sending progress updates.

ProviderMessage

Message enum for ProviderProto

ProviderProto

RequestMode

Request mode for all data related requests.

RequestUpdate

Stream of updates for a single request

ThrottleMode

Throttling mode for requests that support throttling.

Traits

HasErrorCode

Type Aliases

ClientResult

EventResult