iroh_blobs::provider
Trait LazyEvent

pub trait LazyEvent {
    // Required method
    fn call(self) -> Event;
}

Helper to lazyly create an Event, in the case that the event creation is expensive and we want to avoid it if the progress sender is disabled.

Required Methods

fn call(self) -> Event