iroh_blobs::rpc::client
Module tags

Available on crate feature rpc only.

API for tag management.

The purpose of tags is to mark information as important to prevent it from being garbage-collected (if the garbage collector is turned on).

A tag has a name that is an arbitrary byte string. In many cases this will be a valid UTF8 string, but there are also use cases where it is useful to have non string data like integer ids in the tag name.

Tags point to a HashAndFormat.

A tag can point to a hash with format BlobFormat::Raw. In that case it will protect just this blob from being garbage-collected.

It can also point to a hash in format BlobFormat::HashSeq. In that case it will protect the blob itself and all hashes in the blob (the blob must be just a sequence of hashes). Using this format it is possible to protect a large number of blobs with a single tag.

Tags can be created, read, renamed and deleted. Tags do not have to correspond to already existing data. It is perfectly valid to create a tag for data you don’t have yet.

The main entry point is the Client.

Structs

Client

Iroh tags client.

DeleteOptions

Options for a delete operation.

ListOptions

Options for a list operation.

TagInfo

Information about a tag.

Type Aliases

MemClient

A client that uses the memory connector.