iroh_blobs::rpc::proto
Module blobs

Available on crate feature rpc only.

RPC requests and responses for the blob service.

Structs

AddPathRequest

A request to the node to provide the data at the given path

AddPathResponse

Wrapper around AddProgress.

AddStreamRequest

Write a blob from a byte stream

AddStreamResponse

Wrapper around AddProgress.

BatchAddPathRequest

Write a blob from a byte stream

BatchAddPathResponse

Response to a batch add path request

BatchAddStreamRequest

Write a blob from a byte stream

BatchCreateRequest

Request to create a new scope for temp tags

BatchCreateTempTagRequest

Create a temp tag with a given hash and format

BlobStatusRequest

Request to get the status of a blob

BlobStatusResponse

The response to a status request

ConsistencyCheckRequest

A request to the node to validate the integrity of all provided data

CreateCollectionRequest

Create a collection.

CreateCollectionResponse

A response to a create collection request

DeleteRequest

Delete a blob

DownloadResponse

Progress response for BlobDownloadRequest

ExportRequest

A request to the node to download and share the data specified by the hash.

ExportResponse

Progress response for ExportRequest

ListIncompleteRequest

List all blobs, including collections

ListRequest

List all blobs, including collections

ReadAtRequest

Get the bytes for a hash

ValidateRequest

A request to the node to validate the integrity of all provided data

Enums

AddStreamUpdate

Write a blob from a byte stream

BatchAddStreamResponse

Wrapper around AddProgress.

BatchAddStreamUpdate

Write a blob from a byte stream

BatchCreateResponse

Response to a temp tag scope request

BatchUpdate

Update to a temp tag scope

ReadAtResponse

Response to ReadAtRequest

Request

Response