iroh_blobs

Module store

Store implementations

Use the mem store for sharing a small amount of mutable data, the readonly_mem store for sharing static data, and the fs store for when you want to efficiently share more than the available memory and have access to a writeable filesystem.

Modules

fs

File based blob store.

mem

Mutable in-memory blob store.

readonly_mem

Readonly in-memory store.

Constants

IROH_BLOCK_SIZE

Block size used by iroh, 2^4*1024 = 16KiB