Module iroh_blobs::store::bao_file

An implementation of a bao file, meaning some data blob with associated outboard.

Compared to just a pair of (data, outboard), this implementation also works when both the data and the outboard is incomplete, and not even the size is fully known.

There is a full in memory implementation, and an implementation that uses the file system for the data, outboard, and sizes file. There is also a combined implementation that starts in memory and switches to file when the memory limit is reached.

Structs

BaoFileConfig

Configuration for the deferred batch writer. It will start writing to memory, and then switch to a file when the memory limit is reached.

BaoFileHandle

A cheaply cloneable handle to a bao file, including the hash and the configuration.

BaoFileHandleInner

The inner part of a bao file handle.

BaoFileHandleWeak

A weak reference to a bao file handle.

BaoFileWriter

This is finally the thing for which we can implement BaoPairMut.

CompleteStorage

Storage for complete blobs. There is no longer any uncertainty about the size, so we don’t need a sizes file.

DataPaths 🔒

Data files are stored in 3 files. The data file, the outboard file, and a sizes file. The sizes file contains the size that the remote side told us when writing each data block.

DataReader

A reader for a bao file, reading just the data.

FileStorage

A file storage for an incomplete bao file.

OutboardReader

A reader for the outboard part of a bao file.

Enums

BaoFileStorage 🔒

The storage for a bao file. This can be either in memory or on disk.

HandleChange 🔒

Functions

create_read_write 🔒

Create a file for reading and writing, but without truncating the existing file.

max_offset 🔒

read_to_end 🔒

Read from the given file at the given offset, until end of file or max bytes.

with_storage 🔒

Type Aliases

CreateCb 🔒