iroh_blobs::store
Enum ValidateProgress

pub enum ValidateProgress {
    Starting {
        total: u64,
    },
    Entry {
        id: u64,
        hash: Hash,
        path: Option<String>,
        size: u64,
    },
    EntryProgress {
        id: u64,
        offset: u64,
    },
    EntryDone {
        id: u64,
        error: Option<String>,
    },
    PartialEntry {
        id: u64,
        hash: Hash,
        path: Option<String>,
        size: u64,
    },
    PartialEntryProgress {
        id: u64,
        offset: u64,
    },
    PartialEntryDone {
        id: u64,
        ranges: RangeSpec,
    },
    AllDone,
    Abort(Error),
}

Progress updates for the validate operation

Variants

Starting

started validating

Fields

total: u64

The total number of entries to validate

Entry

We started validating a complete entry

Fields

id: u64

a new unique id for this entry

hash: Hash

the hash of the entry

path: Option<String>

location of the entry.

In case of a file, this is the path to the file. Otherwise it might be an url or something else to uniquely identify the entry.

size: u64

The size of the entry, in bytes.

EntryProgress

We got progress ingesting item id.

Fields

id: u64

The unique id of the entry.

offset: u64

The offset of the progress, in bytes.

EntryDone

We are done with id

Fields

id: u64

The unique id of the entry.

error: Option<String>

An error if we failed to validate the entry.

PartialEntry

We started validating an entry

Fields

id: u64

a new unique id for this entry

hash: Hash

the hash of the entry

path: Option<String>

location of the entry.

In case of a file, this is the path to the file. Otherwise it might be an url or something else to uniquely identify the entry.

size: u64

The best known size of the entry, in bytes.

PartialEntryProgress

We got progress ingesting item id.

Fields

id: u64

The unique id of the entry.

offset: u64

The offset of the progress, in bytes.

PartialEntryDone

We are done with id

Fields

id: u64

The unique id of the entry.

ranges: RangeSpec

Available ranges.

AllDone

We are done with the whole operation.

Abort(Error)

We got an error and need to abort.