Enum iroh_blobs::store::fs::ActorMessage

pub(crate) enum ActorMessage {

    Get {
        hash: Hash,
        tx: Sender<Result<Option<BaoFileHandle>, ActorError>>,
    },
    EntryStatus {
        hash: Hash,
        tx: Sender<Result<EntryStatus, ActorError>>,
    },
    GetFullEntryState {
        hash: Hash,
        tx: Sender<Result<Option<EntryData>, ActorError>>,
    },
    SetFullEntryState {
        hash: Hash,
        entry: Option<EntryData>,
        tx: Sender<Result<(), ActorError>>,
    },
    GetOrCreate {
        hash: Hash,
        tx: Sender<Result<BaoFileHandle, ActorError>>,
    },
    OnMemSizeExceeded {
        hash: Hash,
    },
    OnComplete {
        handle: BaoFileHandle,
    },
    Import {
        cmd: Import,
        tx: Sender<Result<(TempTag, u64), ActorError>>,
    },
    Export {
        cmd: Export,
        tx: Sender<Result<(), ActorError>>,
    },
    ImportFlatStore {
        paths: FlatStorePaths,
        tx: Sender<bool>,
    },
    UpdateInlineOptions {
        inline_options: InlineOptions,
        reapply: bool,
        tx: Sender<()>,
    },
    Blobs {
        filter: Box<dyn Fn(u64, AccessGuard<'_, Hash>, AccessGuard<'_, EntryState>) -> Option<(Hash, EntryState)> + Send + Sync>,
        tx: Sender<Result<Vec<Result<(Hash, EntryState), StorageError>>, ActorError>>,
    },
    Tags {
        filter: Box<dyn Fn(u64, AccessGuard<'_, Tag>, AccessGuard<'_, HashAndFormat>) -> Option<(Tag, HashAndFormat)> + Send + Sync>,
        tx: Sender<Result<Vec<Result<(Tag, HashAndFormat), StorageError>>, ActorError>>,
    },
    SetTag {
        tag: Tag,
        value: Option<HashAndFormat>,
        tx: Sender<Result<(), ActorError>>,
    },
    CreateTag {
        hash: HashAndFormat,
        tx: Sender<Result<Tag, ActorError>>,
    },
    Delete {
        hashes: Vec<Hash>,
        tx: Sender<Result<(), ActorError>>,
    },
    Sync {
        tx: Sender<()>,
    },
    Dump,
    Fsck {
        repair: bool,
        progress: BoxedProgressSender<ConsistencyCheckProgress>,
        tx: Sender<Result<(), ActorError>>,
    },
    GcStart {
        tx: Sender<()>,
    },
    Shutdown {
        tx: Option<Sender<()>>,
    },

}

Variants

Get

Fields

hash: Hash

tx: Sender<Result<Option<BaoFileHandle>, ActorError>>

EntryStatus

Query method: get the rough entry status for a hash. Just complete, partial or not found.

Fields

hash: Hash

tx: Sender<Result<EntryStatus, ActorError>>

GetFullEntryState

Query method: get the full entry state for a hash.

Fields

hash: Hash

tx: Sender<Result<Option<EntryData>, ActorError>>

SetFullEntryState

Modification method: set the full entry state for a hash.

Fields

hash: Hash

entry: Option<EntryData>

tx: Sender<Result<(), ActorError>>

GetOrCreate

Modification method: get or create a file handle for a hash.

If the entry exists in redb, either partial or complete, the corresponding data will be returned. If it does not yet exist, a new partial file handle will be created, but not yet written to redb.

Fields

hash: Hash

tx: Sender<Result<BaoFileHandle, ActorError>>

OnMemSizeExceeded

Modification method: inline size was exceeded for a partial entry. If the entry is complete, this is a no-op. If the entry is partial and in memory, it will be written to a file and created in redb.

Fields

hash: Hash

OnComplete

Modification method: marks a partial entry as complete. Calling this on a complete entry is a no-op.

Fields

handle: BaoFileHandle

Import

Modification method: import data into a redb store

At this point the size, hash and outboard must already be known.

Fields

cmd: Import

tx: Sender<Result<(TempTag, u64), ActorError>>

Export

Modification method: export data from a redb store

In most cases this will not modify the store. Only when using ExportMode::TryReference and the entry is large enough to not be inlined.

Fields

cmd: Export

tx: Sender<Result<(), ActorError>>

ImportFlatStore

Modification method: import an entire flat store into the redb store.

Fields

paths: FlatStorePaths

tx: Sender<bool>

UpdateInlineOptions

Update inline options

Fields

inline_options: InlineOptions

The new inline options

reapply: bool

Whether to reapply the new options to existing entries

tx: Sender<()>

Blobs

Bulk query method: get entries from the blobs table

Fields

filter: Box<dyn Fn(u64, AccessGuard<'_, Hash>, AccessGuard<'_, EntryState>) -> Option<(Hash, EntryState)> + Send + Sync>

tx: Sender<Result<Vec<Result<(Hash, EntryState), StorageError>>, ActorError>>

Tags

Bulk query method: get the entire tags table

Fields

filter: Box<dyn Fn(u64, AccessGuard<'_, Tag>, AccessGuard<'_, HashAndFormat>) -> Option<(Tag, HashAndFormat)> + Send + Sync>

tx: Sender<Result<Vec<Result<(Tag, HashAndFormat), StorageError>>, ActorError>>

SetTag

Modification method: set a tag to a value, or remove it.

Fields

tag: Tag

value: Option<HashAndFormat>

tx: Sender<Result<(), ActorError>>

CreateTag

Modification method: create a new unique tag and set it to a value.

Fields

hash: HashAndFormat

tx: Sender<Result<Tag, ActorError>>

Delete

Modification method: unconditional delete the data for a number of hashes

Fields

hashes: Vec<Hash>

tx: Sender<Result<(), ActorError>>

Sync

Sync the entire database to disk.

This just makes sure that there is no write transaction open.

Fields

tx: Sender<()>

Dump

Internal method: dump the entire database to stdout.

Fsck

Internal method: validate the entire database.

Note that this will block the actor until it is done, so don’t use it on a node under load.

Fields

repair: bool

progress: BoxedProgressSender<ConsistencyCheckProgress>

tx: Sender<Result<(), ActorError>>

GcStart

Internal method: notify the actor that a new gc epoch has started.

This will be called periodically and can be used to do misc cleanups.

Fields

tx: Sender<()>

Shutdown

Internal method: shutdown the actor.

Can have an optional oneshot sender to signal when the actor has shut down.

Fields

tx: Option<Sender<()>>

Implementations

impl ActorMessage

fn category(&self) -> MessageCategory