iroh_blobs::store

Module mem

Mutable in-memory blob store.

Being a memory store, this store has to import all data into memory before it can serve it. So the amount of data you can serve is limited by your available memory. Other than that this is a fully featured store that provides all features such as tags and garbage collection.

For many use cases this can be quite useful, since it does not require write access to the file system.

Structs

BaoFileHandle

A cheaply cloneable handle to a bao file, including the hash

BaoFileHandleInner

BaoFileStorageSubscriber

CompleteStorage

DataReader

MemStore

Options

OutboardReader

Enums

BaoFileStorage