iroh_blobs::store
Trait ReadableStore

pub trait ReadableStore: Map {
    // Required methods
    fn blobs(&self) -> impl Future<Output = Result<DbIter<Hash>>> + Send;
    fn tags(
        &self,
        from: Option<Tag>,
        to: Option<Tag>,
    ) -> impl Future<Output = Result<DbIter<(Tag, HashAndFormat)>>> + Send;
    fn temp_tags(
        &self,
    ) -> Box<dyn Iterator<Item = HashAndFormat> + Send + Sync + 'static>;
    fn consistency_check(
        &self,
        repair: bool,
        tx: BoxedProgressSender<ConsistencyCheckProgress>,
    ) -> impl Future<Output = Result<()>> + Send;
    fn partial_blobs(&self) -> impl Future<Output = Result<DbIter<Hash>>> + Send;
    fn export(
        &self,
        hash: Hash,
        target: PathBuf,
        mode: ExportMode,
        progress: ExportProgressCb,
    ) -> impl Future<Output = Result<()>> + Send;
}

Extension of Map to add misc methods used by the rpc calls.

Required Methods

fn blobs(&self) -> impl Future<Output = Result<DbIter<Hash>>> + Send

list all blobs in the database. This includes both raw blobs that have been imported, and hash sequences that have been created internally.

fn tags( &self, from: Option<Tag>, to: Option<Tag>, ) -> impl Future<Output = Result<DbIter<(Tag, HashAndFormat)>>> + Send

list all tags (collections or other explicitly added things) in the database

fn temp_tags( &self, ) -> Box<dyn Iterator<Item = HashAndFormat> + Send + Sync + 'static>

Temp tags

fn consistency_check( &self, repair: bool, tx: BoxedProgressSender<ConsistencyCheckProgress>, ) -> impl Future<Output = Result<()>> + Send

Perform a consistency check on the database

fn partial_blobs(&self) -> impl Future<Output = Result<DbIter<Hash>>> + Send

list partial blobs in the database

fn export( &self, hash: Hash, target: PathBuf, mode: ExportMode, progress: ExportProgressCb, ) -> impl Future<Output = Result<()>> + Send

This trait method extracts a file to a local path.

hash is the hash of the file target is the path to the target file mode is a hint how the file should be exported. progress is a callback that is called with the total number of bytes that have been written

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.