iroh_blobs
Module util

Utility functions and types.

Modules

fs

Utilities for filesystem operations.

io

Utilities for working with tokio io

local_pool

A local task pool with proper shutdown

progress

Utilities for reporting progress.

Structs

SparseMemFile

A file that is sparse in memory

Tag

A tag

TempTag

A hash and format pair that is protected from garbage collection.

Enums

MemOrFile

This is a general purpose Either, just like Result, except that the two cases are Mem for something that is in memory, and File for something that is somewhere external and only available via io.

SetTagOption

Option for commands that allow setting a tag

Traits

TagCounter

A trait for things that can track liveness of blobs and collections.

TagDrop

Trait used from temp tags to notify an abstract store that a temp tag is being dropped.

Functions

total_bytes

Get the number of bytes given a set of chunk ranges and the total size.