iroh_blobs::util::fs
Function key_to_path

pub fn key_to_path(
    key: impl AsRef<[u8]>,
    prefix: Option<String>,
    root: Option<PathBuf>,
) -> Result<PathBuf>

Helper function that translates a key that was derived from the path_to_key function back into a path.

If prefix exists, it will be stripped before converting back to a path If root exists, will add the root as a parent to the created path Removes any null byte that has been appended to the key