iroh_blobs::util
Module progress

Utilities for reporting progress.

The main entry point is the ProgressSender trait.

Structs

AsyncChannelProgressSender

A progress sender that uses an async channel.

BoxedProgressSender

A boxed progress sender

FallibleProgressSliceWriter

A slice writer that adds a fallible progress callback.

IgnoreProgressSender

A no-op progress sender.

ProgressSliceWriter

A slice writer that adds a synchronous progress callback.

WithFilterMap

Transform the message type by filter-mapping to the type of this sender.

WithMap

Transform the message type by mapping to the type of this sender.

Enums

ProgressSendError

An error that can occur when sending progress messages.

Traits

IdGenerator

An id generator, to be combined with a progress sender.

ProgressSender

A general purpose progress sender. This should be usable for reporting progress from both blocking and non-blocking contexts.

Type Aliases

ProgressSendResult

A result type for progress sending.