Trait iroh_blobs::util::progress::BoxableProgressSender

trait BoxableProgressSender<T>: IdGenerator + Debug + Send + Sync + 'static {
    // Required methods
    fn send(
        &self,
        msg: T
    ) -> Pin<Box<dyn Future<Output = ProgressSendResult<()>> + Send + '_>>;
    fn try_send(&self, msg: T) -> ProgressSendResult<()>;
    fn blocking_send(&self, msg: T) -> ProgressSendResult<()>;
}

Boxable progress sender

Required Methods

fn send( &self, msg: T ) -> Pin<Box<dyn Future<Output = ProgressSendResult<()>> + Send + '_>>

Send a message and wait if the receiver is full.

Use this to send important progress messages where delivery must be guaranteed.

fn try_send(&self, msg: T) -> ProgressSendResult<()>

Try to send a message and drop it if the receiver is full.

Use this to send progress messages where delivery is not important, e.g. a self contained progress message.

fn blocking_send(&self, msg: T) -> ProgressSendResult<()>

Send a message and block if the receiver is full.

Use this to send important progress messages where delivery must be guaranteed.