iroh_blobs::util

Trait SendStream

pub trait SendStream: Send {
    // Required methods
    fn send_bytes(
        &mut self,
        bytes: Bytes,
    ) -> impl Future<Output = Result<()>> + Send;
    fn send(&mut self, buf: &[u8]) -> impl Future<Output = Result<()>> + Send;
    fn sync(&mut self) -> impl Future<Output = Result<()>> + Send;
    fn reset(&mut self, code: VarInt) -> Result<()>;
    fn stopped(&mut self) -> impl Future<Output = Result<Option<VarInt>>> + Send;
    fn id(&self) -> u64;
}

An abstract iroh::endpoint::SendStream.

Required Methods

fn send_bytes( &mut self, bytes: Bytes, ) -> impl Future<Output = Result<()>> + Send

Send bytes to the stream. This takes a Bytes because iroh can directly use them.

This method is not cancellation safe. Even if this does not resolve, some bytes may have been written when previously polled.

fn send(&mut self, buf: &[u8]) -> impl Future<Output = Result<()>> + Send

Send that sends a fixed sized buffer.

fn sync(&mut self) -> impl Future<Output = Result<()>> + Send

Sync the stream. Not needed for iroh, but needed for intermediate buffered streams such as compression.

fn reset(&mut self, code: VarInt) -> Result<()>

Reset the stream with the given error code.

fn stopped(&mut self) -> impl Future<Output = Result<Option<VarInt>>> + Send

Wait for the stream to be stopped, returning the error code if it was.

fn id(&self) -> u64

Get the stream id.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.