Trait iroh_blobs::util::TagDrop

pub trait TagDrop: Debug + Send + Sync + 'static {
    // Required method
    fn on_drop(&self, inner: &HashAndFormat);
}

Trait used from temp tags to notify an abstract store that a temp tag is being dropped.

Required Methods

fn on_drop(&self, inner: &HashAndFormat)

Called on drop