Crate iroh_blobs

Blobs layer for iroh.

The crate is designed to be used from the iroh crate, which provides a high level interface, but can also be used standalone.

It implements a protocol for streaming content-addressed data transfer using BLAKE3 verified streaming.

It also provides a store interface for storage of blobs and outboards, as well as a persistent and a memory store implementation.

To implement a server, the provider module provides helpers for handling connections and individual requests given a store.

To perform get requests, the get module provides utilities to perform requests and store the result in a store, as well as a low level state machine for executing requests.

The downloader module provides a component to download blobs from multiple sources and store them in a store.

Re-exports

pub use crate::util::Tag;

pub use crate::util::TempTag;

Modules

downloaderdownloader

Handle downloading blobs and collections concurrently and from nodes.

export

Functions to export data from a store

format

Defines data formats for HashSeq.

get

The client side API

hashseq

traits related to collections of blobs

metrics

Metrics for iroh-blobs

protocol

Protocol for transferring content-addressed blobs and collections over quic connections. This can be used either with normal quic connections when using the quinn crate or with magicsock connections when using the iroh-net crate.

provider

The server side API

store

Implementations of blob stores

util

Utility functions and types.

Structs

Hash

Hash type used throughout.

HashAndFormat

A hash and format pair

Enums

BlobFormat

A format identifier

Constants

IROH_BLOCK_SIZE

Block size used by iroh, 2^4*1024 = 16KiB