iroh_blobs::api

Module blobs

API to interact with a local blob store

This API is for local interactions with the blob store, such as importing and exporting blobs, observing the bitfield of a blob, and deleting blobs.

The main entry point is the Blobs struct.

Structs

AddBytesOptions

Options for adding bytes.

AddPathOptions

Options for adding data from a file system path.

AddProgress

A progress handle for an import operation.

Batch

A batch of operations that modify the blob store.

BatchAddProgress

A progress handle for a batch scoped add operation.

Bitfield

The state of a bitfield, or an update to a bitfield

BlobReader

A reader for blobs that implements AsyncRead and AsyncSeek.

Blobs

Blobs API

BlobsListProgress

A progress handle for a blobs list operation.

DeleteOptions

Options for force deletion of blobs.

ExportBaoOptions

Export the given ranges in bao format, with the iroh block size.

ExportBaoProgress

A progress handle for a bao export operation.

ExportOptions

Export a file to a target path.

ExportProgress

A progress handle for an export operation.

ExportRangesOptions

Export the given ranges as chunkks, without validation.

ExportRangesProgress

A progress handle for a bao export operation.

ImportBaoHandle

A handle for an ongoing bao import operation.

ImportBaoOptions

Import bao encoded data for the given hash with the iroh block size.

ObserveOptions

Observe the local bitfield of the given hash.

ObserveProgress

An observe result. Awaiting this will return the current state.

ReaderOptions

Options for an async reader for blobs that supports AsyncRead and AsyncSeek.

Enums

AddProgressItem

Progress events for importing from any local source.

BlobStatus

Status information about a blob.

EncodedItem

A content item for the bao streaming protocol.

ExportMode

The import mode describes how files will be imported.

ExportProgressItem

Progress events for exporting to a local file.

ImportMode

The import mode describes how files will be imported.