iroh_blobs
Module get

The client side API

To get data, create a connection using iroh-net or use any quinn connection that was obtained in another way.

Create a request describing the data you want to get.

Then create a state machine using fsm::start and drive it to completion by calling next on each state.

For some states you have to provide additional arguments when calling next, or you can choose to finish early.

Modules

db

Functions that use the iroh-blobs protocol in conjunction with a bao store.

error

Error returned from get operations

fsm

Finite state machine for get responses.

progress

Types for get progress state management.

request

Utilities for complex get requests.

Structs

Stats

Stats about the transfer.

Enums

GetResponseError

Error when processing a response