iroh_blobs::get

Module fsm

Finite state machine for get responses.

This is the low level API for getting data from a peer.

Structs

AtBlobContent

State while we are reading content

AtBlobHeader

State before reading a size header

AtClosing

State when finishing the get response

AtConnected

State of the get response machine after the handshake has been sent

AtEndBlob

State after we have read all the content for a blob

AtInitial

Initial state of the get response machine

AtStartChild

State of the get response when we start reading a child

AtStartRoot

State of the get response when we start reading a collection

RequestCounters

Enums

AtBlobHeaderNextError

Error that you can get from AtBlobHeader::next

AtClosingNextError

Error that you can get from AtBlobHeader::next

BlobContentNext

The next state after reading a content item

ConnectedNext

Possible next states after the handshake has been sent

ConnectedNextError

Error that you can get from AtConnected::next

DecodeError

Decode error that you can get once you have sent the request and are decoding the response, e.g. from AtBlobContent::next.

EndBlobNext

The next state after the end of a blob

InitialNextError

Error that you can get from AtConnected::next

Functions

start

The entry point of the get response machine

start_get_many

Start with a get many request. Todo: turn this into distinct states.