iroh_blobs::get
Module request

Utilities for complex get requests.

Functions

get_chunk_probe

Probe for a single chunk of a blob.

get_hash_seq_and_sizes

Given a hash of a hash seq, get the hash seq and the verified sizes of its children.

get_unverified_size

Get the claimed size of a blob from a peer.

get_verified_size

Get the verified size of a blob from a peer.

random_hash_seq_ranges

Given a sequence of sizes of children, generate a range spec that selects a random chunk of a random child.