iroh_blobs
Module store

Implementations of blob stores

Re-exports

pub use bao_tree;

pub use range_collections;

Modules

fsfs-store

redb backed storage

mem

A full in memory database for iroh-blobs

readonly_mem

A readonly in memory database for iroh-blobs, usable for testing and sharing static data.

Structs

GcConfig

Configuration for the GC mark and sweep.

Enums

BaoBlobSize

The size of a bao file

ConsistencyCheckProgress

Progress updates for the validate operation

EntryStatus

The availability status of an entry in a store.

Event

Database events

ExportFormat

The expected format of a hash being exported.

ExportMode

The import mode describes how files will be imported.

ExportProgress

GcMarkEvent

An event related to GC

GcSweepEvent

An event related to GC

ImportMode

The import mode describes how files will be imported.

ImportProgress

Progress messages for an import operation

ReportLevel

Level for generic validation messages

ValidateProgress

Progress updates for the validate operation

Traits

BaoBatchWriter

An async batch interface for writing bao content items to a pair of data and outboard.

Map

A generic map from hashes to bao blobs (blobs with bao outboards).

MapEntry

An entry for one hash in a bao map

MapEntryMut

A partial entry

MapMut

A mutable bao map.

ReadableStore

Extension of Map to add misc methods used by the rpc calls.

Store

The mutable part of a Bao store.

Type Aliases

DbIter

A fallible but owned iterator over the entries in a store.

ExportProgressCb

Export trogress callback