iroh_blobs::store
Trait MapEntry

pub trait MapEntry:
    Debug
    + Clone
    + Send
    + Sync
    + 'static {
    // Required methods
    fn hash(&self) -> Hash;
    fn size(&self) -> BaoBlobSize;
    fn is_complete(&self) -> bool;
    fn outboard(&self) -> impl Future<Output = Result<impl Outboard>> + Send;
    fn data_reader(
        &self,
    ) -> impl Future<Output = Result<impl AsyncSliceReader>> + Send;
}

An entry for one hash in a bao map

The entry has the ability to provide you with an (outboard, data) reader pair. Creating the reader is async and may fail. The futures that create the readers must be Send, but the readers themselves don’t have to be.

Required Methods

fn hash(&self) -> Hash

The hash of the entry.

fn size(&self) -> BaoBlobSize

The size of the entry.

fn is_complete(&self) -> bool

Returns true if the entry is complete.

Note that this does not actually verify if the bytes on disk are complete, it only checks if the entry was marked as complete in the store.

fn outboard(&self) -> impl Future<Output = Result<impl Outboard>> + Send

A future that resolves to a reader that can be used to read the outboard

fn data_reader( &self, ) -> impl Future<Output = Result<impl AsyncSliceReader>> + Send

A future that resolves to a reader that can be used to read the data

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.