iroh_blobs

Module util

Utilities

Modules

connection_pool

A simple iroh connection pool

Structs

AsyncReadRecvStream

AsyncWriteSendStream

Utility to convert a [tokio::io::AsyncWrite] into an SendStream.

RecvStreamAsyncStreamReader

Traits

AsyncReadRecvStreamExtra

This is a helper trait to work with AsyncReadRecvStream. If you have an AsyncRead + Unpin + Send, you can implement these additional methods and wrap the result in an AsyncReadRecvStream to get a RecvStream that reads from the underlying AsyncRead.

AsyncWriteSendStreamExtra

This is a helper trait to work with AsyncWriteSendStream.

RecvStream

An abstract iroh::endpoint::RecvStream.

SendStream

An abstract iroh::endpoint::SendStream.